+++ client/app/locations/locations.component.js
@@ -2,9 +2,13 @@
 import angular from 'angular';
 import routes from './locations.routes';
 import uiRouter from 'angular-ui-router';
+import locations from '../../assets/data/locations.json';
 
 export class LocationsController {
   /*@ngInject*/
+  $onInit() {
+    this.locations = locations;
+  }
 }
 
 export default angular.module('shyApp.locations', [uiRouter])